Fix quote source in renderer
