+++ app/src/main/java/com/devfruit/watermelon/QuoteSource.java
@@ -1,0 +1,14 @@
+package com.devfruit.watermelon;
+
+class QuoteSource {
+
+    String key;
+    String title;
+    String description;
+
+    QuoteSource(String key, String title, String description) {
+        this.key = key;
+        this.title = title;
+        this.description = description;
+    }
+}
