+++ app/src/main/res/values-ru/strings.xml
@@ -5,7 +5,6 @@
         <item>Классика</item>
         <item>Определения</item>
         <item>Английский</item>
-        <item>Фоменко</item>
         <item>Омар Хаям</item>
         <item>Ежи Лец</item>
         <item>Любовь</item>
