+++ app/src/main/res/values/arrays.xml
@@ -4,7 +4,6 @@
         <item>src_classics_biter</item>
         <item>src_definition_biter</item>
         <item>src_english_biter</item>
-        <item>src_fomenko_biter</item>
         <item>src_hayam_biter</item>
         <item>src_lec_biter</item>
         <item>src_love_biter</item>
