+++ app/src/main/res/values/strings.xml
@@ -5,7 +5,6 @@
         <item>Classics</item>
         <item>Definitions</item>
         <item>English</item>
-        <item>Fomenko</item>
         <item>Omar Hayam</item>
         <item>Ezhi Lec</item>
         <item>Love</item>
